complete default contoller

+++ app/Http/Controllers/DefaultsController.php
@@ -12,6 +12,8 @@ class DefaultsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index(Defaults $model) {
+        $this->authorize('manage-defaults', User::class);
+
         return view('default.index', ['defaults' => $model->all()]);
     }
 
@@ -21,32 +23,9 @@ class DefaultsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function create() {
-        return view('default.create');
-    }
+        $this->authorize('manage-defaults', User::class);
 
-    /**
-     * Build the request into the database safe key values.
-     *
-     * @param assoc $attributes The attributes.
-     * @return array The final array.
-     */
-    private function buildFinal($attributes) {
-        $final = ['options' => []];
-        if ($attributes['type'] == 'text')
-            $final['options']['values'] = [];
-        foreach ($attributes as $key => $value) {
-            if (in_array($key, ['_token', '_method', 'name', 'description', 'for_table']))
-                $final[$key] = $value;
-            elseif (preg_match('/^(\w+)_(\d+)$/', $key, $match)) {
-                $i = $match[2];
-                $key = $match[1];
-                if (!isset($final['options']['values'][$i]))
-                    $final['options']['values'][$i] = [];
-                $final['options']['values'][$i][$key] = $value;
-            } else $final['options'][$key] = $value;
-        }
-        $final['options'] = json_encode($final['options']);
-        return $final;
+        return view('default.create');
     }
 
     /**
@@ -56,7 +35,7 @@ class DefaultsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(DefaultRequest $request, Defaults $model) {
-        $model->create($this->buildFinal($request->all()));
+        $model->create($request->all());
         return redirect()->route('default.index')->withStatus(__('Default successfully created.'));
     }
 
@@ -67,7 +46,6 @@ class DefaultsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function edit(Defaults $default) {
-        $default['options'] = json_decode($default['options']);
         return view('default.edit', ['default' => $default]);
     }
 
@@ -79,7 +57,7 @@ class DefaultsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(DefaultRequest $request, Defaults $default) {
-        $default->update($this->buildFinal($request->all()));
+        $default->update($request->all());
         return redirect()->route('default.index')->withStatus(__('Default successfully updated.'));
     }
 
@@ -90,6 +68,8 @@ class DefaultsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy(Defaults $default) {
+        $this->authorize('manage-defaults', User::class);
+
         $default->delete();
         return redirect()->route('default.index')->withStatus(__('Default successfully deleted.'));
     }
